pyproject.toml
[tool.poetry]
name = "mlops-zoomcamp"
version = "0.1.0"
description = ""
authors = ["Shun-Lung Chang <shunlung.chang@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mlops_zoomcamp"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
wget = "^3.2"
mlflow = "^2.3.2"
xgboost = "^1.7.5"
optuna = "^3.1.1"
wandb = "^0.15.3"
prefect = "^2.10.12"
prefect-aws = "^0.3.2"
python-dotenv = "^1.0.0"
prefect-email = "^0.2.2"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.3.0"}
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
